Refactor UserService for DTO usage and exception handling
- Replace `UserDto` with `CreateUserDto` and `UserResponseDto`.
- Integrate `UserMapper` for DTO mapping.
- Add `UserNotFoundException` in user-related methods.
- Update service methods to return DTOs instead of entities.

+++ src/main/java/com/example/task_manager/service/UserService.java
@@ -1,7 +1,10 @@
 package com.example.task_manager.service;
 
-import com.example.task_manager.dto.UserDto;
+import com.example.task_manager.dto.user.CreateUserDto;
+import com.example.task_manager.dto.user.UserResponseDto;
 import com.example.task_manager.entity.User;
+import com.example.task_manager.exception.UserNotFoundException;
+import com.example.task_manager.mapper.UserMapper;
 import com.example.task_manager.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
@@ -10,41 +13,53 @@ import java.util.List;
 @Service
 public class UserService {
 
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
+    private final UserMapper userMapper;
 
-    public User save(UserDto userDto) {
-        User user = this.hydrateUser(userDto);
-        return userRepository.save(user);
+    public UserService(UserRepository userRepository, UserMapper userMapper) {
+        this.userRepository = userRepository;
+        this.userMapper = userMapper;
     }
 
-    public User update(User user) {
-        return userRepository.save(user);
+    public UserResponseDto save(CreateUserDto userDto) {
+        User user = userMapper.toEntity(userDto);
+        User userSave =  userRepository.save(user);
+
+        return userMapper.toResponseDto(userSave);
     }
 
-    public List<User> findAll() {
-        return userRepository.findAll();
+    public UserResponseDto update(Long idUser, CreateUserDto userDto) {
+
+        User userFind = userRepository.findById(idUser).orElseThrow(() -> new UserNotFoundException(idUser));
+
+        userFind.setName(userDto.getName());
+        userFind.setPassword(userDto.getPassword());
+        userFind.setEmail(userDto.getEmail());
+        userFind.setRole(userDto.getRole());
+
+        User userUpdated = userRepository.save(userFind);
+        return userMapper.toResponseDto(userUpdated);
     }
 
-    public User findById(Long id) {
-        return userRepository.findById(id).orElse(null);
+    public List<UserResponseDto> findAll() {
+        return userRepository.findAll().stream().map(userMapper::toResponseDto).toList();
     }
 
-    public User findByEmail(String email) {
-        return userRepository.findByEmail(email).orElse(null);
+    public UserResponseDto findById(Long id) {
+        User userFind = userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
+        return userMapper.toResponseDto(userFind);
+    }
+
+    public UserResponseDto findByEmail(String email) {
+        User userFind = userRepository.findByEmail(email).orElseThrow(() -> new UserNotFoundException(email));
+        return userMapper.toResponseDto(userFind);
     }
 
     public void delete(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
-    }
-
-    private User hydrateUser(UserDto userDto) {
-        User user = new User();
-        user.setName(userDto.getName());
-        user.setRole(userDto.getRole());
-        user.setEmail(userDto.getEmail());
-        user.setPassword(userDto.getPassword());
-
-        return user;
     }
 
 }
